Fixed some GCC compiler warnings and removed deprecated functions

+++ modules/juce_audio_processors/format_types/juce_AU_Shared.h
@@ -308,7 +308,7 @@ struct AudioUnitHelpers
     static bool isLayoutSupported (const AudioProcessor& processor,
                                    bool isInput, int busIdx,
                                    int numChannels,
-                                   const short (&channelLayoutList) [numLayouts][2],
+                                   const short (&channelLayoutList) [(size_t) numLayouts][2],
                                    bool hasLayoutMap = true)
     {
         if (const AudioProcessor::Bus* bus = processor.getBus (isInput, busIdx))
