+++ modules/juce_audio_processors/processors/juce_AudioPluginInstance.h
@@ -116,7 +116,7 @@ protected:
     AudioPluginInstance() = default;
     AudioPluginInstance (const BusesProperties& ioLayouts) : AudioProcessor (ioLayouts) {}
     template <int numLayouts>
-    AudioPluginInstance (const short channelLayoutList[numLayouts][2]) : AudioProcessor (channelLayoutList) {}
+    AudioPluginInstance (const short channelLayoutList[(size_t) numLayouts][2]) : AudioProcessor (channelLayoutList) {}
 
 private:
     void assertOnceOnDeprecatedMethodUse() const noexcept;
